feat: Header 컴포넌트에서 스크롤에 따른 크기 조정 로직 이동 및 최적화 #3
